+++ customise/app/deletePerson.php
@@ -1,8 +1,13 @@
-<?php 
+<?php
 include 'includes/dbConnection.php';
 $toRemove = $_GET['id'];
 
-mysqli_query($con, "DELETE FROM user_favPeople WHERE favePersonID = '$toRemove'");	
+if (substr($_SERVER['HTTP_REFERER'], strrpos($_SERVER['HTTP_REFERER'], '/') + 1) != 'settings.php') {
+	header("Location: index.php");
+	die();
+}
+
+mysqli_query($con, "DELETE FROM user_favPeople WHERE favePersonID = '$toRemove'");
 mysqli_close($con);
 header("Location: settings.php");
 ?>
